# Conflicts:
#	database/migrations/2025_04_07_130758_001_create_collection_product_table.php

+++ database/migrations/2025_04_07_130758_001_create_collection_product_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::create('collection_product', function (Blueprint $table) {
-            $table->unsignedBigInteger('collection_id');
+            $table->integer('collection_id');
             $table->unsignedBigInteger('product_id');
 
             // Define composite primary key
